bool flag probability maps

+++ napari_yapic_prediction/yapic_prediction.py
@@ -1,4 +1,5 @@
 from .yapic_dependencies import NapariSession
+from datetime import datetime
 from skimage import io
 import numpy as np
 import h5py
@@ -7,17 +8,18 @@ import os
 
 def get_lbl_map(model_path):
     """Returns a dictionary retriving the label mapping done by Yapic.
-    
+
     Keys: Napari label values.
     Values: YAPiC label mapping.
     """
     f = h5py.File(model_path, 'r')
     map_dataset = f['lbl_map']
     return {nap: yap for yap, nap in map_dataset}
-    
-    
-def tif_2_layer(tif_path, lbl_map):
-    """Reads temporal tiff file and returns a numpy array. Tiff file is deleted before within the function."""
+
+
+def tif_2_layer(tif_path, lbl_map, save_flag):
+    """Reads temporal tiff file and returns a numpy array. Tiff file is deleted
+    before within the function."""
     img_data = io.imread(tif_path)
     if len(img_data.shape) == 3:
         img_data = np.expand_dims(img_data, axis=0)
@@ -29,13 +31,16 @@ def tif_2_layer(tif_path, lbl_map):
                 pixel_probs = img_data[z, x, y]
                 max_index = np.argmax(pixel_probs)
                 label_output[z, x, y] = lbl_map[max_index + 1]
-    os.remove(tif_path)
+    if not save_flag:
+        os.remove(tif_path)
     return label_output.astype(int)
 
 
-def yapic_prediction(model_path, viewer, progress_map, progress_upload):
-    """This function adds the label layers corresponding to the prediction of all the images uploaded in the Napari viewer.
-    
+def yapic_prediction(model_path, viewer, progress_map,
+                     progress_upload, save_flag):
+    """This function adds the label layers corresponding to the prediction of
+    all the images uploaded in the Napari viewer.
+
     Parameters
         ----------
         model_path : str
@@ -48,35 +53,41 @@ def yapic_prediction(model_path, viewer, progress_map, progress_upload):
             PyQt progress bar instance.
         progress_upload : PyQt5.QtWidgets.QProgressBar
             PyQt progress bar instance.
+        save_flag : int
+            Flag to store probability maps.
     """
-    assert os.path.isfile(model_path) and os.path.splitext(model_path)[-1] == '.h5', '<network> must be a h5 model file'
-    
-    # temporal folder creation
+    model_name = os.path.splitext(model_path)[0]
+    assert os.path.isfile(model_path) and os.path.splitext(
+        model_path)[-1] == '.h5', '<network> must be a h5 model file'
+
+    # folder creation
     work_dir = os.path.dirname(model_path)
-    tmp_dir = os.path.join(work_dir, 'tmp')
-    os.mkdir(tmp_dir)
-    
+    date_time = datetime.now().strftime("%d-%m-%Y_%H-%M-%S")
+    result_dir = os.path.join(
+        work_dir, '{}_results_{}'.format(model_name, date_time))
+    os.mkdir(result_dir)
+
     # Geting label mapping values (dictionary)
     lbl_map = get_lbl_map(model_path)
-    
+
     # predicting
     s = NapariSession()
-    s.load_prediction_data(viewer, tmp_dir)
+    s.load_prediction_data(viewer, result_dir)
     s.load_model(model_path)
     s.set_normalization('local')
     s.predict(True, progress_map)
-    
+
     # Adding to Napari
-    files2upload = os.listdir(tmp_dir)
+    files2upload = os.listdir(result_dir)
     N = len(files2upload)
-    
+
     for i, label_file in enumerate(files2upload):
-        file_path = os.path.join(tmp_dir, label_file)
+        file_path = os.path.join(result_dir, label_file)
         label_name = label_file.split('.')[0]
-        label_data = tif_2_layer(file_path, lbl_map)
+        label_data = tif_2_layer(file_path, lbl_map, save_flag)
         viewer.add_labels(label_data, name='{}_prediction'.format(label_name))
         progress_upload.setValue((i + 1) * 100 / N)
-        
+
     # Deleting temporal folder
-    os.rmdir(tmp_dir)
-    
+    if not save_flag:
+        os.rmdir(result_dir)
